Created a 'Card' ui component
